# validation handler benchmark tests