--- a/pkg/webhook/testdata/psp-all-violations/psp-constraints/host-filesystem-constraint.yaml
+++ b/pkg/webhook/testdata/psp-all-violations/psp-constraints/host-filesystem-constraint.yaml
@@ -1,0 +1,13 @@
+apiVersion: constraints.gatekeeper.sh/v1beta1
+kind: K8sPSPHostFilesystem
+metadata:
+  name: psp-host-filesystem
+spec:
+  match:
+    kinds:
+      - apiGroups: [""]
+        kinds: ["Pod"]
+  parameters:
+    allowedHostPaths:
+    - readOnly: true
+      pathPrefix: "/foo"