--- a/pkg/webhook/testdata/psp-all-violations/psp-constraints/host-namespaces-constraint.yaml
+++ b/pkg/webhook/testdata/psp-all-violations/psp-constraints/host-namespaces-constraint.yaml
@@ -1,0 +1,9 @@
+apiVersion: constraints.gatekeeper.sh/v1beta1
+kind: K8sPSPHostNamespace
+metadata:
+  name: psp-host-namespace
+spec:
+  match:
+    kinds:
+      - apiGroups: [""]
+        kinds: ["Pod"]