--- a/pkg/webhook/testdata/psp-all-violations/psp-constraints/host-network-constraint.yaml
+++ b/pkg/webhook/testdata/psp-all-violations/psp-constraints/host-network-constraint.yaml
@@ -1,0 +1,13 @@
+apiVersion: constraints.gatekeeper.sh/v1beta1
+kind: K8sPSPHostNetworkingPorts
+metadata:
+  name: psp-host-network-ports
+spec:
+  match:
+    kinds:
+      - apiGroups: [""]
+        kinds: ["Pod"]
+  parameters:
+    hostNetwork: true
+    min: 80
+    max: 9000