--- a/pkg/webhook/testdata/psp-all-violations/psp-constraints/privileged-containers-constraint.yaml
+++ b/pkg/webhook/testdata/psp-all-violations/psp-constraints/privileged-containers-constraint.yaml
@@ -1,0 +1,9 @@
+apiVersion: constraints.gatekeeper.sh/v1beta1
+kind: K8sPSPPrivilegedContainer
+metadata:
+  name: psp-privileged-container
+spec:
+  match:
+    kinds:
+      - apiGroups: [""]
+        kinds: ["Pod"]