--- a/pkg/webhook/testdata/psp-all-violations/psp-pods/host-filesystem-example.yaml
+++ b/pkg/webhook/testdata/psp-all-violations/psp-pods/host-filesystem-example.yaml
@@ -1,0 +1,18 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: nginx-host-filesystem
+  labels:
+    app: nginx-host-filesystem
+spec:
+  containers:
+  - name: nginx
+    image: nginx
+    volumeMounts:
+    - mountPath: /cache
+      name: cache-volume
+      readOnly: true
+  volumes:
+  - name: cache-volume
+    hostPath:
+      path: /tmp # directory location on host