--- a/pkg/webhook/testdata/psp-all-violations/psp-pods/host-namespaces-example.yaml
+++ b/pkg/webhook/testdata/psp-all-violations/psp-pods/host-namespaces-example.yaml
@@ -1,0 +1,12 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: nginx-host-namespace
+  labels:
+    app: nginx-host-namespace
+spec:
+  hostPID: true #false
+  hostIPC: true #false
+  containers:
+  - name: nginx
+    image: nginx