--- a/pkg/webhook/testdata/psp-all-violations/psp-pods/host-network-example.yaml
+++ b/pkg/webhook/testdata/psp-all-violations/psp-pods/host-network-example.yaml
@@ -1,0 +1,14 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: nginx-host-networking-ports
+  labels:
+    app: nginx-host-networking-ports
+spec:
+  hostNetwork: true #false
+  containers:
+  - name: nginx
+    image: nginx
+    ports:
+    - containerPort: 9001
+      hostPort: 9001