--- a/pkg/webhook/testdata/psp-all-violations/psp-pods/privileged-containers-example.yaml
+++ b/pkg/webhook/testdata/psp-all-violations/psp-pods/privileged-containers-example.yaml
@@ -1,0 +1,12 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: nginx-privileged
+  labels:
+    app: nginx-privileged
+spec:
+  containers:
+  - name: nginx
+    image: nginx
+    securityContext:
+      privileged: true #false