--- a/pkg/webhook/testdata/psp-all-violations/psp-pods/volumes-example.yaml
+++ b/pkg/webhook/testdata/psp-all-violations/psp-pods/volumes-example.yaml
@@ -1,0 +1,24 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: nginx-volume-types
+  labels:
+    app: nginx-volume-types
+spec:
+  containers:
+  - name: nginx
+    image: nginx
+    volumeMounts:
+    - mountPath: /cache
+      name: cache-volume
+  - name: nginx2
+    image: nginx
+    volumeMounts:
+    - mountPath: /cache2
+      name: demo-vol
+  volumes:
+  - name: cache-volume
+    hostPath:
+      path: /tmp # directory location on host
+  - name: demo-vol
+    emptyDir: {}