**What this PR does / why we need it**:

- The PR performs a benchmark of Gatekeeper's validating handler. 

- A sample of PSP constraint templates are pre-loaded into OPA. Each subtest pre-loads M constraints where M varies from 5 to 100000. For each of the M constraints, number of admission requests that can be processed in 1 second is measured.

**Which issue(s) this PR fixes** *(optional, using `fixes #<issue number>(, fixes #<issue_number>, ...)` format, will close the issue(s) when the PR gets merged)*:
Fixes #

**Special notes for your reviewer**:
- These tests do not measure the HTTP roundtrip to/from kube-api-server

Current results:
cores: 8
mem: 16GB

> goos: darwin
> goarch: amd64
> pkg: github.com/open-policy-agent/gatekeeper/pkg/webhook
> BenchmarkValidationHandler/5-8 	     129	   9385359 ns/op
> 
> BenchmarkValidationHandler/10-8         	      78	  18034209 ns/op
> 
> BenchmarkValidationHandler/50-8         	      15	  78782760 ns/op
> 
> BenchmarkValidationHandler/100-8        	       7	 208199975 ns/op
> 
> BenchmarkValidationHandler/200-8        	       4	 329748445 ns/op
> 
> BenchmarkValidationHandler/1000-8       	       1	1846845599 ns/op
> 
> BenchmarkValidationHandler/2000-8       	       1	3729658819 ns/op
> 
> BenchmarkValidationHandler/10000-8      	       1	17484500470 ns/op
> 
> BenchmarkValidationHandler/20000-8      	       1	35564250359 ns/op
> 
> BenchmarkValidationHandler/100000-8     	       1	165198614593 ns/op

With memory stats

> goos: darwin
> goarch: amd64
> pkg: github.com/open-policy-agent/gatekeeper/pkg/webhook
> 
> BenchmarkValidationHandler/5-8 	     141	   7852219 ns/op	 3658307 B/op	   56034 allocs/op
> 
> BenchmarkValidationHandler/10-8         	      79	  15501142 ns/op	 7221695 B/op	  110534 allocs/op
> 
> BenchmarkValidationHandler/50-8         	      14	  73609200 ns/op	35739040 B/op	  547383 allocs/op
> 
> BenchmarkValidationHandler/100-8        	       7	 148565541 ns/op	71472198 B/op	 1095124 allocs/op
> 
> BenchmarkValidationHandler/200-8        	       4	 305889273 ns/op	143454862 B/op	 2200257 allocs/op
> 
> BenchmarkValidationHandler/1000-8       	       1	1401957112 ns/op	689022800 B/op	10466498 allocs/op
> 
> BenchmarkValidationHandler/2000-8       	       1	2874750865 ns/op	1377815096 B/op	20930929 allocs/op
> 
> BenchmarkValidationHandler/10000-8      	       1	14572899917 ns/op	6888856152 B/op	104646480 allocs/op
> 
> BenchmarkValidationHandler/20000-8      	       1	29800708818 ns/op	13777700688 B/op	209290439 allocs/op
> 
> BenchmarkValidationHandler/100000-8     	       1	161775454897 ns/op	69075147392 B/op	1046442550 allocs/op