@shravan-achar 2020-04-27
@maxsmythe @ritazh @shomron  I have added some tests to measure number of admission requests that can be handled by Gatekeeper when varying number of constraints are pre-loaded into OPA. Please take a look to see if this can be helpful. I have also pasted one set of results in the PR description. 

The output follows the following format:
{{Test Name}}/{{Number of Constraints}}-{{CPU cores}} {{Number of iterations}} {{Average time for each iteration in ns/op}}

For ex: 
`BenchmarkValidationHandler/5-8 129 9385359 ns/op` would mean that a maximum of 129 pod admission requests were reviewed by GK's validation handler in 1 second. Average time for each iteration was 9.3 microseconds with 5 constraints were pre-loaded. 

The tests run atleast 1 iteration even if takes more than 1 second to complete. This is seen for `BenchmarkValidationHandler/1000-8 1 1846845599 ns/op` where 1000 constraints were pre-loaded and a single iteration took 1.8 seconds. 

Please let me know if you have any questions.