@shravan-achar 2020-05-05
@maxsmythe - Thanks for taking a look. I am happy to add more tests in another PR.  
Could you clarify what testing is being referred to here
> When we are doing kind testing in the template itself