@maxsmythe 2020-05-05
re: taking a look, absolutely, happy to look :)

WRT "When we are doing kind testing in the template itself" I am referring to when there is an explicit kind check in the `violation` rule as opposed to using match semantics, e.g.:

```
// only deny Pods named bad
violation[{"msg": "not allowed"}] {
   input.review.kind.kind == "Pod"
   input.review.object.metadata.name == "bad"
}
```