@shravan-achar 2020-05-13
@maxsmythe - I have changed the test based on earlier feedback. Separated out templates, constraints and resources into separate test folders. Modified helper method signatures.

One can curate a list of templates, constraints and objects to define a new test. Non-applicable constraints or violating/non-violating requests could be added as separate test cases. Please give it a look when you get a chance. 
cc: @ritazh @shomron