@maxsmythe 2020-05-16
Thank you for the contribution! I'm glad it was fun :)

Of course we'd love it if you continued to contribute! The more the merrier!

I was definitely interested in the potential for benchmarking the adding of constraint templates and the caching of data. Templates is a known pain point that we should address some day.

Other than that, let us know if there are any issues you're interested in working on or designs you'd like to address and we'll do our best to get you set up.